refactor cancelAlarm back to KeepAliveReceiver
