+++ app/src/main/java/info/nightscout/androidaps/MainApp.java
@@ -70,7 +70,6 @@ import info.nightscout.androidaps.plugins.XDripStatusline.StatuslinePlugin;
 import info.nightscout.androidaps.receivers.DataReceiver;
 import info.nightscout.androidaps.receivers.KeepAliveReceiver;
 import info.nightscout.androidaps.receivers.NSAlarmReceiver;
-import info.nightscout.utils.LocalAlertUtils;
 import info.nightscout.utils.NSUpload;
 import io.fabric.sdk.android.Fabric;
 
@@ -216,7 +215,7 @@ public class MainApp extends Application {
 
     public void stopKeepAliveService() {
         if (keepAliveReceiver != null)
-            LocalAlertUtils.cancelAlarm(this);
+            KeepAliveReceiver.cancelAlarm(this);
     }
 
     public static Bus bus() {
